Combine mixins with command creation (#82)
* Move SystemMixin

* Move group commands

* Move browse commands

* Move player mixin
