+++ pyheos/__init__.py
@@ -21,7 +21,7 @@ from .error import (
     HeosError,
 )
 from .group import HeosGroup
-from .heos import Heos, HeosOptions, PlayerUpdateResult
+from .heos import Heos
 from .media import (
     AlbumMetadata,
     BrowseResult,
@@ -33,12 +33,14 @@ from .media import (
     RetreiveMetadataResult,
     ServiceOption,
 )
+from .options import HeosOptions
 from .player import (
     CONTROLS_ALL,
     CONTROLS_FORWARD_ONLY,
     CONTROLS_PLAY_STOP,
     HeosNowPlayingMedia,
     HeosPlayer,
+    PlayerUpdateResult,
     PlayMode,
 )
 from .search import MultiSearchResult, SearchCriteria, SearchResult, SearchStatistic
